% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_session.R
\name{connect_session}
\alias{connect_session}
\title{Connect a session to a PhantomJS process}
\usage{
connect_session(url)
}
\arguments{
\item{url}{string indicating the URL you wish to use}
}
\value{
a list with phantom JS process and object of class \code{Session}
}
\description{
Connect a session to a PhantomJS process
}
