% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_root_url.R
\name{create_root_url}
\alias{create_root_url}
\title{Creates a root URL form an inputted URL}
\usage{
create_root_url(input.URL)
}
\arguments{
\item{input.URL}{Specified URL string}
}
\value{
Root of imputted URL
}
\description{
Creates the root URL of a given URL. Function can only take a object of length 1 in order to work
}
