% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table.R
\name{create_table}
\alias{create_table}
\title{Create a data frame of cleaned URLs}
\usage{
create_table(input)
}
\arguments{
\item{input}{String or vector of URLs. Must have a length greater than 1.}
}
\description{
Creates a data frame from a vector of inputted URLs that provides the cleaned, root, and redirected versions of the inputted URLs
}
