% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_link.R
\name{download_link}
\alias{download_link}
\title{Download data from a link}
\usage{
download_link(pjs_session, xpath, save_dir = "temp", save_as = NULL)
}
\arguments{
\item{pjs_session}{phantom JS session object}

\item{xpath}{string indicating the xpath of the download data link button}

\item{save_dir}{string indicating the download directory - defaults to tempdir}

\item{save_as}{string indicating the save file name - defaults to temp with the extension provided by the download link}
}
\description{
Download data from a link
}
