% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_proof.r
\name{error_proof}
\alias{error_proof}
\title{Run functions safely}
\usage{
error_proof(x, return_na = FALSE)
}
\arguments{
\item{x}{A function.}

\item{return_na}{Logical, true to return NA when function fails, FALSE to return nothing.}
}
\value{
The output \code{x} when \code{x} runs succesfully. When \code{x} is unsuccessful, returns NA when \code{return_na} is TRUE, otherwise NULL..
}
\description{
Run functions safely
}
\examples{
error_proof(mean("a"), return_na=TRUE)
}
