% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_html.R
\name{fetch_html}
\alias{fetch_html}
\title{Fetch HTML content from a URL}
\usage{
fetch_html(url)
}
\arguments{
\item{url}{The URL to fetch HTML content from.}
}
\value{
The fetched HTML content as a character string.
}
\description{
This function fetches HTML content from a given URL using the `httr` package.
}
