% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_links.R
\name{get_links}
\alias{get_links}
\title{Get all links on a scraped HTML page}
\usage{
get_links(input.URL)
}
\arguments{
\item{input.URL}{Normalized URL that can be scraped.}
}
\value{
Vector of all URLs listed on the inputted webpage
}
\description{
Get all links on a scraped HTML page
}
