% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_p_node_data.R
\name{get_p_node_data}
\alias{get_p_node_data}
\title{Create a data frame that contains information about p nodes in every internal links in a given URL}
\usage{
get_p_node_data(input.URL)
}
\arguments{
\item{input.URL}{An URL for searching every p nodes in its interal links}
}
\value{
A data frame
}
\description{
First, this function finds redirected url from a given url. This redirected url is used for
subsequent parts of this function. It makes a list of internal links using LinkExtractor function
in Rcrawler package. Then, information related to every p nodes in each internal link is scraped and
combined in a data frame. This information include xpath, text, url, domain, and tag.
}
\examples{
input.URL <- "HTTP://GMFD.ORG/GMFRA/GMFRAINDEX.HTM"
get_p_node_data( input.URL )
}
