% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_redirected_url.R
\name{get_redirected_url}
\alias{get_redirected_url}
\title{Get a redirected URL from an input URL}
\usage{
get_redirected_url(input.URL)
}
\arguments{
\item{input.URL}{An URL}
}
\value{
An URL redirected
}
\description{
This function returns a redirected URL. If the website of input URL is contained within a frame,
it extract the URL embedded within the frame and use that URL as the redirect value.
}
\examples{
input.URL <- "HTTP://GMFD.ORG/GMFRA/GMFRAINDEX.HTM"
get_redirected_url( input.URL )
}
