% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_UK_weather_warnings.r
\name{get_UK_weather_warnings}
\alias{get_UK_weather_warnings}
\title{Parse the met office RSS feed for weather warnings into a tidy data frame.
"http://www.metoffice.gov.uk/public/data/PWSCache/WarningsRSS/Region/UK"}
\usage{
get_UK_weather_warnings()
}
\value{
Returns a tidy dataframe of all UK weather warnings disaggregated by region..
}
\description{
Parse the met office RSS feed for weather warnings into a tidy data frame.
"http://www.metoffice.gov.uk/public/data/PWSCache/WarningsRSS/Region/UK"
}
\examples{
get_UK_weather_warnings()
}
