% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url_formats.R
\name{get_url_formats}
\alias{get_url_formats}
\title{Create a data frame of different URL formats}
\usage{
get_url_formats(input.URL)
}
\arguments{
\item{input.URL}{Single URL with working URL}
}
\value{
Table of different URL formats the inputted URL
}
\description{
Create a data frame of different URL formats
}
