% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_test_urls.R
\name{load_test_urls}
\alias{load_test_urls}
\title{Load sample test URLs from the 1023-EZ forms}
\usage{
load_test_urls()
}
\value{
A data frame of test URLs
}
\description{
This functions loads in 20 test URLs from the 1023-EZ forms. These URLs are a sample of the entire URLs provided
intended to check the functionality of the the other URLs cleaning and scraping functions in the package.
}
