% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_html.R
\name{parse_html}
\alias{parse_html}
\title{Parse HTML content}
\usage{
parse_html(html_content)
}
\arguments{
\item{html_content}{The HTML content to parse.}
}
\value{
A parsed HTML document object.
}
\description{
This function parses HTML content using the `rvest` package.
}
