% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_chart_ts.R
\name{scrape_chart_ts}
\alias{scrape_chart_ts}
\title{scrape timeseries data from highcharts}
\usage{
scrape_chart_ts(pjs_session)
}
\arguments{
\item{pjs_session}{phantom JS session object}
}
\value{
\code{data.frame} containing date and value
}
\description{
scrape timeseries data from highcharts
}
