% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_commodities.r
\name{scrape_commodities}
\alias{scrape_commodities}
\title{Scrape historical commodity prices}
\usage{
scrape_commodities(URL)
}
\arguments{
\item{URL}{A url from "https://uk.investing.com/commodities/".}
}
\value{
A dataframe of historical prices for the selected commodity
}
\description{
Scrape historical commodity prices
}
