% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_exchange_rates.r
\name{scrape_exchange_rates}
\alias{scrape_exchange_rates}
\title{Scrape historical exchange rates from exchangerates.org.uk}
\usage{
scrape_exchange_rates(URL)
}
\arguments{
\item{URL}{A url from ""https://www.exchangerates.org.uk/".}
}
\value{
A dataframe of historical exchange rates for the selected currency.
}
\description{
Scrape historical exchange rates from exchangerates.org.uk
}
