% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_ft_company.r
\name{scrape_ft_company}
\alias{scrape_ft_company}
\title{Scrape share prices from a particular company from the financial times.}
\usage{
scrape_ft_company(url, pause = 0)
}
\arguments{
\item{url}{A url from "https://markets.ft.com/data/equities/tearsheet/".}

\item{pause}{Indicates how long the function should sleep for before returning a value.}
}
\value{
A dataframe of historical share statistics from a particular company. Use \code{pause} when vectorising this function
over a list of urls to minimise nuisance to target website.
}
\description{
Scrape share prices from a particular company from the financial times.
}
