% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_lse_company.r
\name{scrape_lse_company}
\alias{scrape_lse_company}
\title{Scrape share prices from a particular company from the London Stock Exchange website.}
\usage{
scrape_lse_company(url, pause = 0)
}
\arguments{
\item{url}{A url from "https://www.londonstockexchange.com".}

\item{pause}{Indicates how long the function should sleep for before returning a value.}
}
\value{
A dataframe of historical share statistics from a particular company. Use \code{pause} when vectorising this function
over a list of urls to minimise nuisance to target website.
}
\description{
Scrape share prices from a particular company from the London Stock Exchange website.
}
\examples{
scrape_lse_company(
url=
"https://www.londonstockexchange.com/exchange/prices-and-markets/stocks/summary/company-summary/GB0008847096GBGBXSET1.html"
)

scrape_lse_company(
url=
"https://www.londonstockexchange.com/exchange/prices-and-markets/stocks/summary/company-summary/GB0009697037GBGBXSTMM.html"
)

scrape_lse_company(
url=
"https://www.londonstockexchange.com/exchange/prices-and-markets/stocks/summary/company-summary/GB0006731235GBGBXSET1.html",
pause=2
)
}
