% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_mysupermarket_prices.r
\name{scrape_mysupermarket_prices}
\alias{scrape_mysupermarket_prices}
\title{Scrape supermarket prices from mysupermarket.co.uk}
\usage{
scrape_mysupermarket_prices(url, pause = 0)
}
\arguments{
\item{url}{A search recult url from "https://www.mysupermarket.co.uk/".}

\item{pause}{Indicates how long the function should sleep for before returning a value.}
}
\value{
A dataframe of prices, pack sizes, and product ids. Use \code{pause} when vectorising this function
over a list of urls to minimise nuisance to target website.
}
\description{
Scrape supermarket prices from mysupermarket.co.uk
}
\examples{
scrape_mysupermarket_prices(url="https://www.mysupermarket.co.uk/Shopping/FindProducts.aspx?Query=cadburys\%20chocolate")
scrape_mysupermarket_prices(url="https://www.mysupermarket.co.uk/Shopping/FindProducts.aspx?query=baked+beans&store=Tesco")
scrape_mysupermarket_prices(url="https://www.mysupermarket.co.uk/Shopping/FindProducts.aspx?query=ice+cream&store=Tesco&_fcategory=Ice_Lollies")
}
