% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_shipais.r
\name{scrape_shipais}
\alias{scrape_shipais}
\title{Scrape ship data from "http://www.shipais.com/"}
\usage{
scrape_shipais(url, pause = 0)
}
\arguments{
\item{url}{A search recult url from "http://www.shipais.com/"}

\item{pause}{Indicates how long the function should sleep for before returning a value.}
}
\value{
A dataframe of ship name, speed, status, destination, ETA and details. Use \code{pause} when vectorising this function
over a list of urls to minimise nuisance to target website.
}
\description{
Scrape ship data from "http://www.shipais.com/"
}
