% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_table.R
\name{scrape_table}
\alias{scrape_table}
\title{scrape tables from html nodes}
\usage{
scrape_table(pjs_session, xpath)
}
\arguments{
\item{pjs_session}{phantom JS session object}

\item{xpath}{string indicating the xpath of the table you wish to scrape}
}
\value{
scraped table data as a \code{data.frame}
}
\description{
scrape tables from html nodes
}
