% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_vesseltracker.r
\name{scrape_vesseltracker}
\alias{scrape_vesseltracker}
\title{Scrape ship data from "https://www.vesseltracker.com/"}
\usage{
scrape_vesseltracker(url, pause = 0)
}
\arguments{
\item{url}{A search recult url from "https://www.vesseltracker.com/"}

\item{pause}{Indicates how long the function should sleep for before returning a value.}
}
\value{
A list containing two dataframes, PortStatus contians general information about the port, PortCalls contains a list of recent arrivals. Use \code{pause} when vectorising this function
over a list of urls to minimise nuisance to target website.
}
\description{
Scrape ship data from "https://www.vesseltracker.com/"
}
\examples{
scrape_vesseltracker(url="https://www.vesseltracker.com/en/Port/Dover/Dashboard.html")
scrape_vesseltracker(url="https://www.vesseltracker.com/en/Port/Calais/Dashboard.html")
scrape_vesseltracker(url="https://www.vesseltracker.com/en/Port/Belfast/Dashboard.html")
}
