% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webscrapeR.r
\docType{package}
\name{webscrapeR}
\alias{webscrapeR}
\alias{webscrapeR-package}
\title{webscrapeR: A package for webscraping various useful data sources..}
\description{
The webscrapeR package contains a collection of functions for parsing useful data out of several websites.
The sources of data include: Supermarket prices, share prices, financial market price history, commoditiy prices.
}
\section{Foo functions}{

scrape_mysupermarket()
error_proof()
}

