% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_financial_market_history.R
\name{scrape_financial_market_history}
\alias{scrape_financial_market_history}
\title{Scrape historical values for different financial markets}
\usage{
scrape_financial_market_history(url)
}
\arguments{
\item{url}{A url from "https://www.sharesmagazine.co.uk/indices/index/".}
}
\value{
A dataframe of historical values for the selected market.
}
\description{
Scrape historical values for different financial markets
}
\examples{
scrape_financial_market_history(
url="https://www.sharesmagazine.co.uk/indices/index/IXIC/historic-prices"
)

scrape_financial_market_history(
url="https://www.sharesmagazine.co.uk/indices/index/UKX/historic-prices"
)

scrape_financial_market_history(
url="https://www.sharesmagazine.co.uk/indices/index/MCX/historic-prices"
)
}
